% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sim.R
\name{analyze_sim}
\alias{analyze_sim}
\title{Analyze data simulated with the \code{sim_scm} function}
\usage{
analyze_sim(data)
}
\arguments{
\item{data}{An object with simulated data of type {.cls ThesisSimData} to perform analyses on.}
}
\value{
A list with the original data and results of the analyses.
}
\description{
Data are analyzed with the RI-CLPM and the DPM (with and without free loadings).
}
\examples{
\dontrun{
analyze_sim(data)
}
}
