% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_scm.R
\name{sim_scm}
\alias{sim_scm}
\title{Simulate data from a Structural Causal Model (SCM) used in the research report}
\usage{
sim_scm(
  timepoints = 5,
  burnin = 45,
  N = 500,
  ndat = 1,
  phi,
  betac,
  betac2 = NULL,
  time_beta_change = NULL,
  psi,
  intercepts = c(0, 0),
  meanc = c(0, 0),
  varC = c(1, 1),
  seed = NULL
)
}
\arguments{
\item{timepoints}{Numeric input specifying the number of timepoints to sample. Used for naming the variables.}

\item{burnin}{Numeric input specifying the number of burnin timepoints to use for convergence. Used for naming the variables. Together with timepoints determines the total number of timepoints sampled.}

\item{N}{Numeric input specifying the sample size.}

\item{ndat}{Numeric input specifying the number of datasets to simulate.}

\item{phi}{2x2 matrix specifying the lagged effects (i.e. autoregressions on the diagonals and cross-lagged effects on the off-diagonals).}

\item{betac}{2x2 matrix specifying the initial effect of the confounders.}

\item{betac2}{Optional 2x2 matrix specifying the changed effects of the confounders.}

\item{time_beta_change}{Optional numeric input specifying the timepoint at which the effects of the confounders change.}

\item{psi}{Atomic vector with length=2 specifying the residual variances of the variables.}

\item{intercepts}{Atomic vector with length=2 specifying the intercepts.}

\item{meanc}{Atomic vector with length=2 specifying the means of the confounders.}

\item{varC}{Atomic vector with length=2 specifying the variances of the confounders.}

\item{seed}{Optional argument specifying the seed to use before resampling.}
}
\value{
An object of class {.cls ThesisSimData}. It is a list that includes the parameters set for the analysis and a ndat x 1 dataset with ndat nested simulated datasets, each of size N x (timepoints+burnin+3).
}
\description{
Simulate data from a Structural Causal Model (SCM) used in the research report. It uses 2 confounders and allows for one change in confounder effects.
}
\examples{
\dontrun{
# set phi matrix
phixx <- 0.2
phixy <- 0.15
phiyy <- 0.30
phiyx <- 0.10
phi <- matrix(c(phixx, phixy,
               phiyx, phiyy),
             nrow = 2,
             byrow=T)
# speciy residual variances
psix <- 1
psiy <- 1
psixy <- c(psix, psiy)
# specify variances of confounders
psic1 <- 1
psic2 <- 1
psic <- c(psic1, psic2)
# specify effects of confounders
betaxc1 <- 0.3
betaxc2 <- 0.8

betayc1 <- 0.5
betayc2 <- 0.2

# put in matrix
betac_1 <- matrix(c(betaxc1, betaxc2,
                   betayc1, betayc2),
                 nrow = 2,
                 byrow = T)

data <- sim_scm(
 timepoints = 5,
 burnin = 5,
 N = 500,
 ndat = 10,
 phi = phi,
 betac = betac_1,
 betac2 = NULLL,
 time_beta_change = NULL,
 psi = psixy,
 intercepts = c(0,0),
 meanc = c(0,0),
 varC = c(0,0),
 seed = 42)
}
}
